Updating to 1.5 (LAST UPDATE)
Improved speed of the bot.
After this update, the bot will not updated, and the project archived, the bot is full working, so don't need other work on it.

+++ gamivo.py
@@ -7,12 +7,13 @@ from unshortenit import UnshortenIt
 import logging
 import os
 import re
+link = False
 unshortener = UnshortenIt()
 short_url_list = [""]
 short_bitly_list = [""]
 bitly_list = [""]
 long_url_list = [""]
-ref = "" #Put your ref in ""
+ref = "?glv=example" #Replace example with your tag
 shortner = bitly_api.Connection(access_token="") #Put your token in ""
 updater = Updater(token="", use_context=True) #Put your token in ""
 
@@ -29,8 +30,6 @@ def github(update, context):
 def devgroup(update, context):
     context.bot.send_message(chat_id=update.effective_chat.id, text=" @angolodidiegogruppo")
 def devchannel(update, context):
-    context.bot.send_message(chat_id=update.effective_chat.id, text=" @angolodidiego")
-def ref_link(update, context):
     context.bot.send_message(chat_id=update.effective_chat.id, text=" @angolodidiego")
 def send(update, context):
     user_says = " ".join(context.args)
@@ -52,7 +51,7 @@ def status(update, context):
 def text(update, context):
     user = update.message.from_user
     text_message = update.message.text
-    lowered_text_message = text_message.lower()
+    lowered_text_message = text_message
     if "gamivo.com" in lowered_text_message:
         x = re.findall(r"(?:https?://)?(?:www.)?gamivo.com\S*", lowered_text_message)
         if len(x) > 70:
@@ -82,6 +81,7 @@ def text(update, context):
         for y in x:
             short_url = unshortener.unshorten(y)
             if "gamivo.com" in short_url:
+                link = True
                 bitly_list.append(y)
                 if "https://" in short_url:
                     short_url = short_url.split("?")[0] + ref
@@ -94,13 +94,15 @@ def text(update, context):
                     short_url = "https://" + short_url
                     short_url = short_url.split("?")[0] + ref
                     short_url = shortner.shorten(short_url)
-                for i,j in zip(short_bitly_list, bitly_list):
-                    lowered_text_message = lowered_text_message.replace(j,i)
+                short_bitly_list.append(short_url["url"])
+                if "gamivo.com" in short_url["url"]:
+                    for i,j in zip(short_bitly_list, bitly_list):
+                        lowered_text_message = lowered_text_message.replace(j,i)
                 long_url_list.clear()
                 short_url_list.clear()
-        context.bot.send_message(chat_id=update.effective_chat.id, text="@" + user["username"] + " aveva scritto: " + lowered_text_message)
-        context.bot.delete_message(chat_id=update.effective_chat.id, message_id=update.message.message_id)
-                
+        if link is True:
+            context.bot.send_message(chat_id=update.effective_chat.id, text="@" + user["username"] + " aveva scritto: " + lowered_text_message)
+            context.bot.delete_message(chat_id=update.effective_chat.id, message_id=update.message.message_id)
 dispatcher = updater.dispatcher
 start_handler = CommandHandler("start", start)
 info_handler = CommandHandler("about", about)
@@ -108,7 +110,6 @@ version_handler = CommandHandler("version", version)
 github_handler = CommandHandler("github", github)
 devgroup_handler = CommandHandler("devgroup", devgroup)
 devchannel_handler = CommandHandler("devchannel", devchannel)
-ref_link_handler = CommandHandler("ref", ref)
 send_handler = CommandHandler("talk", send)
 status_handler = CommandHandler("status", status)
 text_handler = MessageHandler(Filters.text, text)
@@ -118,7 +119,6 @@ dispatcher.add_handler(version_handler)
 dispatcher.add_handler(github_handler)
 dispatcher.add_handler(devgroup_handler)
 dispatcher.add_handler(devchannel_handler)
-dispatcher.add_handler(ref_link_handler)
 dispatcher.add_handler(send_handler)
 dispatcher.add_handler(status_handler)
 dispatcher.add_handler(text_handler)
